+++ packages/waku/tests/create-pages.test.ts
@@ -4,7 +4,6 @@ import type { TypeEqual } from 'ts-expect';
 import { assert, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { MockedFunction } from 'vitest';
 import { getErrorInfo } from '../src/lib/utils/custom-errors.js';
-import { parsePathWithSlug } from '../src/lib/utils/path.js';
 import { Children } from '../src/minimal/client.js';
 import type { PathsForPages } from '../src/router/base-types.js';
 import type { GetSlugs } from '../src/router/create-pages-utils/inferred-path-types.js';
@@ -26,6 +25,7 @@ import type {
   StaticSlugRoutePathsTuple,
 } from '../src/router/create-pages.js';
 import { unstable_defineRouter } from '../src/router/define-router.js';
+import { parsePathWithSlug } from '../src/router/isomorphic-utils/path-spec.js';
 
 function Fake() {
   return null;
