+++ packages/waku/tests/path.test.ts
@@ -1,11 +1,10 @@
 import { describe, expect, test } from 'vitest';
+import { extname, removeBase } from '../src/lib/utils/path.js';
+import { path2regexp } from '../src/router/define-router-utils/path-spec.js';
 import {
-  extname,
   getPathMapping,
   parsePathWithSlug,
-  path2regexp,
-  removeBase,
-} from '../src/lib/utils/path.js';
+} from '../src/router/isomorphic-utils/path-spec.js';
 
 function matchPath(path: string, input: string) {
   return new RegExp(path2regexp(parsePathWithSlug(path))).test(input);
